+++ SpotifyClone/SpotifyClone/Authorization/Request/TokenRequest.swift
@@ -6,34 +6,29 @@
 //
 
 import Foundation
+import CoreFoundation
 
 struct Token: Decodable {
     let accessToken: String
-}
-
-struct TokenRequestBody: HTTPBodyEncodable {
-    let grantType: String = "authorization_code"
-    let code: String
-    let redirectUri: String
-    let clientId: String
-    let codeVerifier: String
+    let tokenType: String
+    let expiresIn: TimeInterval
+    let refreshToken: String
+    let scope: String
 }
 
 struct TokenRequest: Request {
     var path: String = "/api/token"
     var parameters: [String: Any]
-    var body: HTTPBodyEncodable?
     var headers: [String: String] = [:]
     var method: HTTPMethod = .POST
     
     init(code: String, codeVerifier: String, redirectUri: String) {
         parameters = [
             "code": code,
-            "redirect_uri": "g-spotify://g-spotify-callback",
-            "client_id": "d5266d28457741dc81d9b8172ff32bd5",
+            "redirect_uri": GlobalConstants.redirectURI,
+            "client_id": GlobalConstants.spotifyClientID,
             "grant_type": "authorization_code",
             "code_verifier": codeVerifier
         ]
-        self.body = TokenRequestBody(code: code, redirectUri: redirectUri, clientId: "d5266d28457741dc81d9b8172ff32bd5", codeVerifier: codeVerifier)
     }
 }
