+++ SpotifyClone/SpotifyClone/Common/GlobalConstants.swift
@@ -1,0 +1,13 @@
+//
+//  GlobalConstants.swift
+//  SpotifyClone
+//
+//  Created by Gamal Kubeyev on 06.04.2023.
+//
+
+import Foundation
+
+struct GlobalConstants {
+    static let redirectURI = "g-spotify://g-spotify-callback"
+    static let spotifyClientID = "d5266d28457741dc81d9b8172ff32bd5"
+}
